# shared/contracts/openapi/rpml-web-app.v1.yaml
openapi: 3.1.0
info:
  title: RPML Backend
  description: 'RPML Web App HTTP API (contract snapshot: shared/contracts/openapi/rpml-web-app.v1.yaml).'
  version: 1.0.0
paths:
...
  /api/v1/optimization/instances:
    get:
      tags:
      - optimization
      summary: Каталог dataset-инстансов
      description: Возвращает все доступные `.dat` инстансы из bundled Rios-Solis
        dataset.
      operationId: get_dataset_instance_catalog_api_v1_optimization_instances_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetInstanceCatalogResponse'
        '401':
          description: Пользователь не аутентифицирован.
      security:
      - HTTPBearer: []
  /api/v1/optimization/mc-config/defaults:
    get:
      tags:
      - optimization
      summary: Дефолтные параметры Monte Carlo
      description: Возвращает дефолтную конфигурацию модели стохастических траекторий
        дохода.
      operationId: get_monte_carlo_defaults_api_v1_optimization_mc_config_defaults_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationMonteCarloConfig'
        '401':
          description: Пользователь не аутентифицирован.
      security:
      - HTTPBearer: []
  /api/v1/optimization/run:
    post:
      tags:
      - optimization
      summary: Синхронный запуск оптимизации
      description: Выполняет оптимизацию в рамках одного HTTP-запроса. Возвращает
        базовый план и, при `mc_income=true`, дополнительный Monte Carlo summary.
      operationId: run_optimization_sync_api_v1_optimization_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizationRunRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationRunResponse'
        '400':
          description: Bad Request
          $ref: '#/components/responses/ErrorContent'
        '401':
          description: Пользователь не аутентифицирован.
        '422':
          description: Unprocessable Content
          $ref: '#/components/responses/ErrorContent'
      security:
      - HTTPBearer: []
  /api/v1/optimization/tasks:
    post:
      tags:
      - optimization
      summary: Создать асинхронную задачу оптимизации
      description: Ставит расчет в очередь Celery и сразу возвращает task_id.
      operationId: create_optimization_task_api_v1_optimization_tasks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOptimizationTaskRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOptimizationTaskResponse'
        '401':
          description: Пользователь не аутентифицирован.
        '422':
          description: Unprocessable Content
          $ref: '#/components/responses/ErrorContent'
      security:
      - HTTPBearer: []
  /api/v1/optimization/tasks/{task_id}:
    get:
      tags:
      - optimization
      summary: Статус асинхронной задачи оптимизации
      description: Возвращает текущий статус задачи. При completed в ответе присутствует
        `plan_id`, который можно запросить через `/optimization/plans/{plan_id}`.
      operationId: get_optimization_task_status_api_v1_optimization_tasks__task_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationTaskStatusResponse'
        '401':
          description: Пользователь не аутентифицирован.
        '404':
          $ref: '#/components/responses/ErrorContent'
          description: Not Found
        '422':
          $ref: '#/components/responses/ErrorContent'
          description: Unprocessable Content
  /api/v1/optimization/plans/{plan_id}:
    get:
      tags:
      - optimization
      summary: Получить асинхронный план по ID
      description: Возвращает результат расчета, ранее завершенного через `/optimization/tasks`.
      operationId: get_optimization_plan_api_v1_optimization_plans__plan_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationPlanResponse'
        '401':
          description: Пользователь не аутентифицирован.
        '404':
          $ref: '#/components/responses/ErrorContent'
          description: Not Found
        '422':
          $ref: '#/components/responses/ErrorContent'
          description: Unprocessable Content
components:
  schemas:
...
    CreateOptimizationTaskRequest:
      properties:
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          description: Источник входных данных для фоновой задачи.
          default: scenario_snapshot
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя `.dat` инстанса. Обязательно для `dataset_instance`.
        horizon_months:
          anyOf:
          - type: integer
            maximum: 360.0
            minimum: 1.0
          - type: 'null'
          title: Horizon Months
          description: 'Горизонт планирования. Для dataset mode можно не передавать:
            будет использован горизонт инстанса.'
        ru_mode:
          type: boolean
          title: Ru Mode
          description: Если true, применяются правила RU-режима.
          default: true
        mc_income:
          type: boolean
          title: Mc Income
          description: Если true, в фоне будет рассчитан Monte Carlo summary.
          default: false
        mc_config:
          anyOf:
          - $ref: '#/components/schemas/OptimizationMonteCarloConfig'
          - type: 'null'
          description: Параметры Monte Carlo для фонового расчета.
      type: object
      title: CreateOptimizationTaskRequest
    CreateOptimizationTaskResponse:
      properties:
        task_id:
          type: string
          title: Task Id
          description: Идентификатор асинхронной задачи.
        status:
          type: string
          const: pending
          title: Status
          default: pending
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          description: Режим входных данных задачи.
        horizon_months:
          type: integer
          title: Horizon Months
          description: Фактический горизонт задачи.
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя dataset-инстанса, если задача создана в `dataset_instance`
            режиме.
        ru_mode:
          type: boolean
          title: Ru Mode
          description: RU-режим, зафиксированный в задаче.
        mc_income:
          type: boolean
          title: Mc Income
          description: MC-режим, зафиксированный в задаче.
      type: object
      required:
      - task_id
      - input_mode
      - horizon_months
      - ru_mode
      - mc_income
      title: CreateOptimizationTaskResponse
...
    DatasetInstanceCatalogResponse:
      properties:
        total:
          type: integer
          title: Total
          description: Сколько инстансов доступно в каталоге.
        items:
          items:
            $ref: '#/components/schemas/DatasetInstanceSummaryResponse'
          type: array
          title: Items
      type: object
      required:
      - total
      title: DatasetInstanceCatalogResponse
...
    OptimizationMonteCarloConfig:
      properties:
        n_scenarios:
          type: integer
          minimum: 1.0
          title: N Scenarios
          description: Количество сценариев Monte Carlo.
          default: 16
        seed:
          type: integer
          title: Seed
          description: Базовый seed генератора случайных траекторий.
          default: 42
        rho:
          type: number
          maximum: 0.999
          minimum: -0.999
          title: Rho
          description: AR(1) коэффициент автокорреляции дохода.
          default: 0.55
        sigma:
          type: number
          minimum: 0.0
          title: Sigma
          description: Волатильность логнормального шума.
          default: 0.15
        shock_prob:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Shock Prob
          description: Вероятность шокового снижения дохода в месяце.
          default: 0.04
        shock_severity_mean:
          type: number
          minimum: 0.0
          title: Shock Severity Mean
          description: Средняя глубина шока дохода.
          default: 0.3
        shock_severity_std:
          type: number
          minimum: 0.0
          title: Shock Severity Std
          description: Стандартное отклонение глубины шока дохода.
          default: 0.1
        min_income_floor:
          type: number
          minimum: 0.0
          title: Min Income Floor
          description: Нижняя отсечка месячного дохода.
          default: 1.0
      type: object
      title: OptimizationMonteCarloConfig
...
    OptimizationRunRequest:
      properties:
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          description: 'Источник входных данных: snapshot пользователя или dataset
            instance.'
          default: scenario_snapshot
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя `.dat` инстанса. Обязательно для `dataset_instance`.
        horizon_months:
          anyOf:
          - type: integer
            maximum: 360.0
            minimum: 1.0
          - type: 'null'
          title: Horizon Months
          description: Горизонт планирования в месяцах. Для `scenario_snapshot` обязателен;
            для `dataset_instance` должен совпадать с горизонтом инстанса, если передан.
        ru_mode:
          type: boolean
          title: Ru Mode
          description: Если true, применяются правила RU-режима по досрочному погашению.
          default: true
        mc_income:
          type: boolean
          title: Mc Income
          description: Если true, дополнительно рассчитывается Monte Carlo summary
            по траекториям дохода.
          default: false
        mc_config:
          anyOf:
          - $ref: '#/components/schemas/OptimizationMonteCarloConfig'
          - type: 'null'
          description: Тонкие параметры Monte Carlo. Используются только при `mc_income=true`.
      type: object
      title: OptimizationRunRequest
    OptimizationRunResponse:
      properties:
        status:
          type: string
          title: Status
          description: Статус решения MILP (например, OPTIMAL/FEASIBLE).
        total_cost:
          type: number
          title: Total Cost
          description: Стоимость базового детерминированного плана.
        debts:
          items:
            $ref: '#/components/schemas/OptimizationDebtSummary'
          type: array
          title: Debts
          description: Сводка по долгам, соответствующая строкам матриц платежей и
            остатков.
        payments_matrix:
          items:
            items:
              type: number
            type: array
          type: array
          title: Payments Matrix
          description: Матрица платежей [loan][month].
        balances_matrix:
          items:
            items:
              type: number
            type: array
          type: array
          title: Balances Matrix
          description: Матрица остатков [loan][month].
        savings_vector:
          items:
            type: number
          type: array
          title: Savings Vector
          description: Расчетный остаток доступного бюджета по месяцам для данного
            плана (непотраченные средства, перенесенные вперед).
        horizon_months:
          type: integer
          title: Horizon Months
          description: Использованный горизонт планирования.
        baseline_comparison:
          additionalProperties: true
          type: object
          title: Baseline Comparison
          description: Сравнение MILP-плана с baseline-стратегиями (avalanche/snowball).
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          default: scenario_snapshot
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя dataset-инстанса, если расчет выполнен в `dataset_instance`
            режиме.
        assumptions:
          items:
            type: string
          type: array
          title: Assumptions
          description: Список допущений, использованных при расчете.
        ru_mode:
          type: boolean
          title: Ru Mode
          description: Фактически использованный RU-режим.
        mc_income:
          type: boolean
          title: Mc Income
          description: Флаг Monte Carlo режима для этого расчета.
        mc_summary:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Mc Summary
          description: Агрегаты Monte Carlo по стоимости и времени решения (mean/median/p90,
            доля infeasible и т.д.).
        mc_config:
          anyOf:
          - $ref: '#/components/schemas/OptimizationMonteCarloConfig'
          - type: 'null'
          description: Конфигурация Monte Carlo, использованная для расчета.
        budget_policy:
          type: string
          title: Budget Policy
          description: Политика применения месячного бюджета в модели.
        budget_trace:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Budget Trace
          description: 'Помесячная трассировка бюджета: доход месяца, перенос, доступный
            бюджет, фактические платежи и остаток переноса.'
      type: object
      required:
      - status
      - total_cost
      - payments_matrix
      - balances_matrix
      - horizon_months
      - baseline_comparison
      - ru_mode
      - mc_income
      - budget_policy
      title: OptimizationRunResponse
    OptimizationTaskStatusResponse:
      properties:
        status:
          type: string
          enum:
          - pending
          - completed
          - failed
          title: Status
        task_id:
          type: string
          title: Task Id
          description: Идентификатор асинхронной задачи.
        plan_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Plan Id
          description: Идентификатор готового плана (заполняется при completed).
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          description: Текст ошибки (заполняется при failed).
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          description: Режим входных данных задачи.
        horizon_months:
          type: integer
          title: Horizon Months
          description: Горизонт задачи.
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя dataset-инстанса, если задача считает dataset instance.
        ru_mode:
          type: boolean
          title: Ru Mode
          description: RU-режим задачи.
        mc_income:
          type: boolean
          title: Mc Income
          description: MC-режим задачи.
      type: object
      required:
      - status
      - task_id
      - input_mode
      - horizon_months
      - ru_mode
      - mc_income
      title: OptimizationTaskStatusResponse
...
    ErrorResponse:
      type: object
      required:
      - detail
      title: ErrorResponse
      properties:
        detail:
          description: 'Error detail: string from HTTPException or list for validation
            errors.'
          oneOf:
          - type: string
          - type: array
            items: {}
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
    ErrorContent:
      description: Error body for many 4xx/5xx responses (HTTPException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'